+++ cocos2dx/tileMap_parallax_nodes/CCTMXLayer.cpp
@@ -397,7 +397,7 @@ namespace cocos2d {
 
 		CCAssert( item, "TMX atlas index not found. Shall not happen");
 
-        int index = ((size_t)item - (size_t)m_pAtlasIndexArray->arr) / sizeof(void*);
+        int index = (int)((size_t)item - (size_t)m_pAtlasIndexArray->arr) / sizeof(void*);
 		return index;
 	}
 	unsigned int CCTMXLayer::atlasIndexForNewZ(int z)
@@ -406,7 +406,7 @@ namespace cocos2d {
 		unsigned int i=0;
 		for( i=0; i< m_pAtlasIndexArray->num ; i++) 
 		{
-            int val = (size_t) m_pAtlasIndexArray->arr[i];
+            int val = (int)(size_t) m_pAtlasIndexArray->arr[i];
 			if( z < val )
 				break;
 		}	
@@ -473,7 +473,7 @@ namespace cocos2d {
 		CCAssert( m_pChildren->containsObject(sprite), "Tile does not belong to TMXLayer");
 
 		unsigned int atlasIndex = sprite->getAtlasIndex();
-        unsigned int zz = (size_t) m_pAtlasIndexArray->arr[atlasIndex];
+        unsigned int zz = (int)(size_t) m_pAtlasIndexArray->arr[atlasIndex];
 		m_pTiles[zz] = 0;
 		ccCArrayRemoveValueAtIndex(m_pAtlasIndexArray, atlasIndex);
 		CCSpriteBatchNode::removeChild(sprite, cleanup);
